Added MetaAccessible interface back to pointer and member data pointer

+++ tests/unittest/include/dataprovider.h
@@ -53,7 +53,7 @@ public:
 	}
 
 	int getDataCount() const {
-		return dataList.size();
+		return (int)dataList.size();
 	}
 
 	T getData(const int index) const {
